+++ src/impex.js
@@ -46,8 +46,8 @@
 	     * @property {function} toString 返回版本
 	     */
 		this.version = {
-	        v:[0,9,3],
-	        state:'beta',
+	        v:[0,9,4],
+	        state:'',
 	        toString:function(){
 	            return impex.version.v.join('.') + ' ' + impex.version.state;
 	        }